requirements.txt
# Core Flask
flask>=3.0.0,<4

# Supabase integration
supabase==2.20.0

# Environment variables
python-dotenv==1.1.1

# JWT handling
PyJWT==2.10.1

# Additional utilities
requests>=2.31.0
